requirements.txt
langchain
langchain-community
langchain-groq
langchain-text-splitters
langchain-core
pypdf
faiss-cpu
sentence-transformers
streamlit
python-dotenv